Refactor creating packages for classes in the auth package
