+++ src/main/java/store/wckd/server/auth/userdetailservice/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package store.wckd.server.auth;
+package store.wckd.server.auth.userdetailservice;
 
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
